requirements.txt
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
boto3>=1.26.0
python-dotenv>=1.0.0
